Test getDurationSinceLastBreak returns correct duration
This commit had the purpose to uncomment an existing test that would check if the method getDurationSinceLastBreak from the HydrateReminderPlugin class would return the correct duration.
To make this test work, I added a second parameter to the method which represents the current Instant. This way it was possible in the test to mock two instants so that they would always have a duration of 1 hour between them.
Due to the change in the method parameters, the code where this method was used was refactored to reflect this changes without changing how the code worked previously