Finished the sound effect task
This is my last commit for this task. I added a line to import SoundEffectID and a method called playSoundEffect() to play the sound. This method is moddled after the playAnimationForLocalPlayer() method, and gets called in the incrementCurrentSessionHydrationBreaks() method. The modified code passsed all the tests and seemed to work in game.