Sound effect config setting (#404)
* Added a config section for sound effects called Sound Effect Settings in HydrateReminderConfig.java

* Added a checkbox for enabling sound effects caled 'Hydrate sound effect' in HydrateReminderConfig.java

* Finished the sound effect task

This is my last commit for this task. I added a line to import SoundEffectID and a method called playSoundEffect() to play the sound. This method is moddled after the playAnimationForLocalPlayer() method, and gets called in the incrementCurrentSessionHydrationBreaks() method. The modified code passsed all the tests and seemed to work in game.

* Delete .DS_Store

* Remove unwanted files

---------

Co-authored-by: jmakhack <joshmak@berkeley.edu>